+++ backend/canisters/proposals_bot/impl/src/lifecycle/post_upgrade.rs
@@ -1,6 +1,6 @@
 use crate::lifecycle::{init_env, init_state};
 use crate::memory::get_upgrades_memory;
-use crate::{mutate_state, Data};
+use crate::Data;
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
 use ic_cdk_macros::post_upgrade;
@@ -24,15 +24,4 @@ fn post_upgrade(args: Args) {
     init_state(env, data, args.wasm_version);
 
     info!(version = %args.wasm_version, "Post-upgrade complete");
-
-    mutate_state(|state| {
-        for (governance_canister_id, proposal_id) in state.data.nervous_systems.get_finished_user_submitted_proposals() {
-            state
-                .data
-                .finished_proposals_to_process
-                .push_back((governance_canister_id, proposal_id));
-        }
-
-        crate::jobs::start(state);
-    });
 }